refactor the source tree
