+++ intern_dart_project/test/intern_dart_project_test.dart
@@ -6,8 +6,4 @@ void main() {
 }
 
 void testForUnitTest1() {
-  test('testForUnitTest1', () {
-    final expectedList = [1, 2, 3, 4, 5, 6, 10, 12, 15, 20, 25, 30, 50, 60, 75, 100, 150, 300];
-    expect(expectedList, findDivisorOf(300));
-  });
 }
